+++ navis/core/skeleton.py
@@ -394,6 +394,13 @@ class TreeNeuron(BaseNeuron):
                                            rename=True,
                                            optional={('radius', 'W'): 0},
                                            restrict=False)
+
+        # Make sure we don't end up with object dtype anywhere as this can
+        # cause problems
+        for c in ('node_id', 'parent_id'):
+            if self._nodes[c].dtype == 'O':
+                self._nodes[c] = self._nodes[c].astype(int)
+
         graph.classify_nodes(self)
 
     @property
@@ -458,7 +465,7 @@ class TreeNeuron(BaseNeuron):
     @property
     @requires_nodes
     def cycles(self) -> Optional[List[int]]:
-        """Cycles in neuron if any.
+        """Cycles in neuron (if any).
 
         See also
         --------
@@ -477,11 +484,7 @@ class TreeNeuron(BaseNeuron):
 
     @property
     def simple(self) -> 'TreeNeuron':
-        """Return simple neuron representation.
-
-        Consists only of root, branch points and leafs.
-
-        """
+        """Simplified representation consisting only of root, branch points and leafs."""
         if not hasattr(self, '_simple'):
             self._simple = self.downsample(float('inf'),
                                            inplace=False)
@@ -594,6 +597,12 @@ class TreeNeuron(BaseNeuron):
         """Number of branch points."""
         return self.nodes[self.nodes.type == 'branch'].shape[0]
 
+    @property
+    @requires_nodes
+    def n_leafs(self) -> Optional[int]:
+        """Number of leaf nodes."""
+        return self.nodes[self.nodes.type == 'end'].shape[0]
+
     @temp_property
     def cable_length(self) -> Union[int, float]:
         """Cable length."""
@@ -675,7 +684,7 @@ class TreeNeuron(BaseNeuron):
 
     @property
     def sampling_resolution(self) -> float:
-        """Average cable length between 2 nodes."""
+        """Average cable length between child -> parent nodes."""
         return self.cable_length / self.n_nodes
 
     @temp_property
@@ -710,7 +719,7 @@ class TreeNeuron(BaseNeuron):
 
     @property
     def n_skeletons(self) -> int:
-        """Return number of seperate skeletons in this neuron."""
+        """Number of seperate skeletons in this neuron."""
         return len(self.root)
 
     def _clear_temp_attr(self, exclude: list = []) -> None: