+++ navis/io/nmx_io.py
@@ -222,6 +222,11 @@ def read_nmx(f: Union[str, pd.DataFrame, Iterable],
     -------
     navis.NeuronList
 
+    See Also
+    --------
+    :func:`navis.read_nml`
+                        Read NML file(s).
+
     """
     reader = NMXReader(precision=precision,
                        attrs=kwargs)
@@ -279,6 +284,11 @@ def read_nml(f: Union[str, pd.DataFrame, Iterable],
     Returns
     -------
     navis.NeuronList
+
+    See Also
+    --------
+    :func:`navis.read_nmx`
+                        Read NMX files (collections of NML files).
 
     """
     reader = NMLReader(precision=precision,